Moved the class files into an independent package
