+++ jme3-android/src/main/java/com/jme3/app/jmeSurfaceView/OnExceptionThrown.java
@@ -29,21 +29,16 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.jme3.app;
-
-import com.jme3.system.AppSettings;
+package com.jme3.app.jmeSurfaceView;
 
 /**
- * Embedded interface class(abstract) to listen for the moment when when the GL thread holding the #{@link JmeSurfaceView}
- * joins the UI thread , after asynchronous rendering.
- * @see JmeSurfaceView#setOnRendererCompleted(OnRendererCompleted).
+ * Embedded interface designed to listen to exceptions & fire when an exception is thrown.
+ * @see JmeSurfaceView#setOnExceptionThrown(OnExceptionThrown).
  */
-public interface OnRendererCompleted {
+public interface OnExceptionThrown {
     /**
-     * Listens for the the moment when when the GL thread holding the #{@link JmeSurfaceView}
-     * joins the UI thread , after asynchronous rendering.
-     * @param application the current jme game instance.
-     * @param appSettings the current window settings of the running jme game.
+     * Listens for a thrown exception or a thrown error.
+     * @param e the exception or the error that is throwable.
      */
-    void onRenderCompletion(LegacyApplication application , AppSettings appSettings);
+    void onExceptionThrown(Throwable e);
 }
